perf(logging): reduce peak memory usage in CDF export
On a synthetic log file with 100K lines this change brought the total memory usage down 50% from 460 MB to 227 MB (plus 85 MB baseline of just libraries) and, as a side effect, brought the time down 37% from 25.7s to 16.1s.
